moved env.py content into utils/env.py; backward compatibility ref
