+++ libs/langchain/langchain/env.py
@@ -1,17 +1,4 @@
-import platform
-from functools import lru_cache
+"""Keep here for backwards compatibility."""
+from langchain.utils.env import get_runtime_environment
 
-
-@lru_cache(maxsize=1)
-def get_runtime_environment() -> dict:
-    """Get information about the environment."""
-    # Lazy import to avoid circular imports
-    from langchain import __version__
-
-    return {
-        "library_version": __version__,
-        "library": "langchain",
-        "platform": platform.platform(),
-        "runtime": "python",
-        "runtime_version": platform.python_version(),
-    }
+__all__ = ["get_runtime_environment"]
